[FIX] Querys data error

+++ querys/3/reducer.py
@@ -20,17 +20,17 @@ for line in sys.stdin:
     thisKey, thisClose, thisDate = data_mapped
 
     if oldKey and oldKey != thisKey:
-        print("%s\t%s\t%s" % (oldKey, max, int(date)))
+        print("%s\t%s\t%s" % (oldKey, max, date))
         oldKey = thisKey
         max = 0
         date = 0
 
     oldKey = thisKey
-    auxDate = datetime.fromtimestamp(float(thisDate))
+    auxDate = datetime.fromtimestamp(int(thisDate))
 
     if auxDate > thirty_days_ago and max < float(thisClose):
         max = float(thisClose)
         date = auxDate
 
 if oldKey != None:
-   print("%s\t%s\t%s" % (oldKey, max, int(date)))
+   print("%s\t%s\t%s" % (oldKey, max, date))
